refactor(auction-house): refactor to add auctioneer handlers and scopes
