+++ auction-house/program/src/deposit/mod.rs
@@ -1,0 +1,240 @@
+use anchor_lang::{prelude::*, solana_program::program::invoke, AnchorDeserialize};
+
+use crate::{constants::*, errors::*, utils::*, AuctionHouse, AuthorityScope, *};
+
+/// Accounts for the [`deposit` handler](auction_house/fn.deposit.html).
+#[derive(Accounts)]
+#[instruction(escrow_payment_bump: u8)]
+pub struct InstantDeposit<'info> {
+    /// User wallet account.
+    pub wallet: Signer<'info>,
+
+    /// User SOL or SPL account to transfer funds from.
+    #[account(mut)]
+    pub payment_account: UncheckedAccount<'info>,
+
+    /// SPL token account transfer authority.
+    pub transfer_authority: UncheckedAccount<'info>,
+
+    /// Buyer escrow payment account PDA.
+    #[account(mut, seeds=[PREFIX.as_bytes(), auction_house.key().as_ref(), wallet.key().as_ref()], bump=escrow_payment_bump)]
+    pub escrow_payment_account: UncheckedAccount<'info>,
+
+    /// Auction House instance treasury mint account.
+    pub treasury_mint: Account<'info, Mint>,
+
+    /// Auction House instance authority account.
+    pub authority: UncheckedAccount<'info>,
+
+    /// Auction House instance PDA account.
+    #[account(seeds=[PREFIX.as_bytes(), auction_house.creator.as_ref(), auction_house.treasury_mint.as_ref()], bump=auction_house.bump, has_one=authority, has_one=treasury_mint, has_one=auction_house_fee_account)]
+    pub auction_house: Account<'info, AuctionHouse>,
+
+    /// Auction House instance fee account.
+    #[account(mut, seeds=[PREFIX.as_bytes(), auction_house.key().as_ref(), FEE_PAYER.as_bytes()], bump=auction_house.fee_payer_bump)]
+    pub auction_house_fee_account: UncheckedAccount<'info>,
+
+    pub token_program: Program<'info, Token>,
+    pub system_program: Program<'info, System>,
+    pub rent: Sysvar<'info, Rent>,
+}
+
+impl<'info> From<DepositWithAuctioneer<'info>> for InstantDeposit<'info> {
+    fn from(a: DepositWithAuctioneer<'info>) -> InstantDeposit<'info> {
+        InstantDeposit {
+            wallet: a.wallet,
+            payment_account: a.payment_account,
+            transfer_authority: a.transfer_authority,
+            escrow_payment_account: a.escrow_payment_account,
+            treasury_mint: a.treasury_mint,
+            authority: a.authority,
+            auction_house: a.auction_house,
+            auction_house_fee_account: a.auction_house_fee_account,
+            token_program: a.token_program,
+            system_program: a.system_program,
+            rent: a.rent,
+        }
+    }
+}
+
+/// Accounts for the [`deposit` handler](auction_house/fn.deposit.html).
+#[derive(Accounts, Clone)]
+#[instruction(escrow_payment_bump: u8)]
+pub struct DepositWithAuctioneer<'info> {
+    /// User wallet account.
+    pub wallet: Signer<'info>,
+
+    /// User SOL or SPL account to transfer funds from.
+    #[account(mut)]
+    pub payment_account: UncheckedAccount<'info>,
+
+    /// SPL token account transfer authority.
+    pub transfer_authority: UncheckedAccount<'info>,
+
+    /// Buyer escrow payment account PDA.
+    #[account(mut, seeds=[PREFIX.as_bytes(), auction_house.key().as_ref(), wallet.key().as_ref()], bump=escrow_payment_bump)]
+    pub escrow_payment_account: UncheckedAccount<'info>,
+
+    /// Auction House instance treasury mint account.
+    pub treasury_mint: Account<'info, Mint>,
+
+    /// Auction House instance authority account.
+    pub authority: UncheckedAccount<'info>,
+
+    /// Auction House instance PDA account.
+    #[account(seeds=[PREFIX.as_bytes(), auction_house.creator.as_ref(), auction_house.treasury_mint.as_ref()], bump=auction_house.bump, has_one=authority, has_one=treasury_mint, has_one=auction_house_fee_account)]
+    pub auction_house: Account<'info, AuctionHouse>,
+
+    /// Auction House instance fee account.
+    #[account(mut, seeds=[PREFIX.as_bytes(), auction_house.key().as_ref(), FEE_PAYER.as_bytes()], bump=auction_house.fee_payer_bump)]
+    pub auction_house_fee_account: UncheckedAccount<'info>,
+
+    /// The auctioneer program PDA running this auction.
+    pub auctioneer_authority: UncheckedAccount<'info>,
+
+    /// The auctioneer PDA owned by Auction House storing scopes.
+    #[account(seeds = [SALE_AUTHORITY.as_bytes(), auction_house.key().as_ref(), auctioneer_authority.key().as_ref()], bump = auction_house.bump)]
+    pub ah_auctioneer_pda: UncheckedAccount<'info>,
+
+    pub token_program: Program<'info, Token>,
+    pub system_program: Program<'info, System>,
+    pub rent: Sysvar<'info, Rent>,
+}
+
+pub fn instant_deposit<'info>(
+    mut ctx: Context<'_, '_, '_, 'info, InstantDeposit<'info>>,
+    escrow_payment_bump: u8,
+    amount: u64,
+) -> ProgramResult {
+    let auction_house = &ctx.accounts.auction_house;
+
+    // If it has an auctioneer authority delegated must use *_with_auctioneer handler.
+    if auction_house.has_auctioneer {
+        return Err(ErrorCode::MustUseAuctioneerHandler.into());
+    }
+
+    deposit(&mut ctx.accounts, escrow_payment_bump, amount)
+}
+
+pub fn deposit_with_auctioneer<'info>(
+    ctx: Context<'_, '_, '_, 'info, DepositWithAuctioneer<'info>>,
+    escrow_payment_bump: u8,
+    amount: u64,
+) -> ProgramResult {
+    let auction_house = &ctx.accounts.auction_house;
+    let auctioneer_authority = &ctx.accounts.auctioneer_authority;
+    let ah_auctioneer_pda = &ctx.accounts.ah_auctioneer_pda;
+
+    if !auction_house.has_auctioneer {
+        return Err(ErrorCode::NoAuctioneerProgramSet.into());
+    }
+
+    assert_valid_auctioneer_and_scope(
+        &auction_house.key(),
+        &auctioneer_authority.key(),
+        ah_auctioneer_pda,
+        AuthorityScope::Sell,
+    )?;
+
+    let mut accounts: InstantDeposit<'info> = (*ctx.accounts).clone().into();
+
+    deposit(&mut accounts, escrow_payment_bump, amount)
+}
+
+fn deposit<'info>(
+    accounts: &mut InstantDeposit<'info>,
+    escrow_payment_bump: u8,
+    amount: u64,
+) -> ProgramResult {
+    let wallet = &accounts.wallet;
+    let payment_account = &accounts.payment_account;
+    let transfer_authority = &accounts.transfer_authority;
+    let escrow_payment_account = &accounts.escrow_payment_account;
+    let authority = &accounts.authority;
+    let auction_house = &accounts.auction_house;
+    let auction_house_fee_account = &accounts.auction_house_fee_account;
+    let treasury_mint = &accounts.treasury_mint;
+    let system_program = &accounts.system_program;
+    let token_program = &accounts.token_program;
+    let rent = &accounts.rent;
+
+    // If it has an auctioneer authority delegated must use *_with_auctioneer handler.
+    if auction_house.has_auctioneer {
+        return Err(ErrorCode::MustUseAuctioneerHandler.into());
+    }
+
+    let auction_house_key = auction_house.key();
+    let seeds = [
+        PREFIX.as_bytes(),
+        auction_house_key.as_ref(),
+        FEE_PAYER.as_bytes(),
+        &[auction_house.fee_payer_bump],
+    ];
+    let wallet_key = wallet.key();
+
+    let escrow_signer_seeds = [
+        PREFIX.as_bytes(),
+        auction_house_key.as_ref(),
+        wallet_key.as_ref(),
+        &[escrow_payment_bump],
+    ];
+
+    let (fee_payer, fee_seeds) = get_fee_payer(
+        authority,
+        auction_house,
+        wallet.to_account_info(),
+        auction_house_fee_account.to_account_info(),
+        &seeds,
+    )?;
+
+    let is_native = treasury_mint.key() == spl_token::native_mint::id();
+
+    create_program_token_account_if_not_present(
+        escrow_payment_account,
+        system_program,
+        &fee_payer,
+        token_program,
+        treasury_mint,
+        &auction_house.to_account_info(),
+        rent,
+        &escrow_signer_seeds,
+        fee_seeds,
+        is_native,
+    )?;
+
+    if !is_native {
+        assert_is_ata(payment_account, &wallet.key(), &treasury_mint.key())?;
+        invoke(
+            &spl_token::instruction::transfer(
+                token_program.key,
+                &payment_account.key(),
+                &escrow_payment_account.key(),
+                &transfer_authority.key(),
+                &[],
+                amount,
+            )?,
+            &[
+                escrow_payment_account.to_account_info(),
+                payment_account.to_account_info(),
+                token_program.to_account_info(),
+                transfer_authority.to_account_info(),
+            ],
+        )?;
+    } else {
+        assert_keys_equal(payment_account.key(), wallet.key())?;
+        invoke(
+            &system_instruction::transfer(
+                &payment_account.key(),
+                &escrow_payment_account.key(),
+                amount,
+            ),
+            &[
+                escrow_payment_account.to_account_info(),
+                payment_account.to_account_info(),
+                system_program.to_account_info(),
+            ],
+        )?;
+    }
+
+    Ok(())
+}
