+++ auction-house/program/src/state.rs
@@ -1,13 +1,8 @@
 use anchor_lang::{prelude::*, AnchorDeserialize, AnchorSerialize};
 
-pub mod buy;
-pub mod sell;
-pub use buy::*;
-pub use sell::*;
-
 #[derive(AnchorSerialize, AnchorDeserialize, Clone, Default)]
 pub struct Auctioneer {
-    pub auctioneer_program: Pubkey,
+    pub auctioneer_authority: Pubkey,
     pub auction_house: Pubkey,
     pub scopes: Vec<AuthorityScope>,
 }
